# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: HireX Unified Candidate + Recruiter API
  version: 1.0.0-draft
  description: 'DRAFT API v1 contract for Candidate Android and Recruiter Web. HTTP Method + normalized Path is
    the unique operation key. JSON fields use camelCase; all IDs are opaque strings; all date-time values are ISO-8601
    UTC and end with Z. OpenAPI is the implementation source of truth. MVP decisions: SGD integer major units; one
    Resume per Candidate; Recruiter registration creates a company; fixed transition matrices; deferred features
    are marked with x-mvp-scope.'
servers:
- url: http://localhost:8080/api/v1
  description: MVP local Web and backend development
- url: http://10.0.2.2:8080/api/v1
  description: MVP Android Emulator access to the host backend
...
security:
- bearerAuth: []
paths:
...
  /candidate/onboarding:
    post:
      tags: [Profile, Resume, Recommendations]
      summary: Atomically complete Candidate profile, default resume, and job preferences
      operationId: completeCandidateOnboarding
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CandidateOnboardingRequest' }
      responses:
        '204': { description: Onboarding completed. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
      x-permission: Candidate self
      x-clients: [Candidate]
...
  /community/posts/{postId}/direct-conversation:
    post:
      tags: [Community]
      summary: Create or reuse an isolated Community direct conversation with the post author
      operationId: startCommunityDirectConversation
      parameters: [{$ref: '#/components/parameters/CommunityPostId'}]
      responses:
        '201': {description: Community direct conversation, content: {application/json: {schema: {$ref: '#/components/schemas/CommunityDirectConversationResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /community/direct-conversations/{conversationId}:
    get:
      tags: [Community]
      summary: Get a Community direct conversation visible only to its participants
      operationId: getCommunityDirectConversation
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      responses:
        '200': {description: Community direct conversation, content: {application/json: {schema: {$ref: '#/components/schemas/CommunityDirectConversationResponse'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /community/direct-conversations/{conversationId}/messages:
    get:
      tags: [Community]
      summary: List messages visible only to the two Community conversation participants
      operationId: listCommunityDirectMessages
      parameters:
      - $ref: '#/components/parameters/ConversationId'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200': {description: Messages, content: {application/json: {schema: {$ref: '#/components/schemas/CommunityDirectMessageListResponse'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    post:
      tags: [Community]
      summary: Send a Community direct message
      operationId: sendCommunityDirectMessage
      parameters: [{$ref: '#/components/parameters/ConversationId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateCommunityCommentRequest'}}}
      responses:
        '201': {description: Message sent}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    CommunityPostId:
      name: postId
      in: path
      required: true
      description: Opaque Community Post identifier.
      schema:
        type: string
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal has the wrong role or insufficient capability
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    CommunityAuthor:
      type: object
      additionalProperties: false
      required: [userId, fullName, avatarUrl, role, companyName]
      properties:
        userId: { type: string }
        fullName: { type: string }
        avatarUrl:
          oneOf:
          - { type: string, format: uri }
          - { type: 'null' }
        role:
          type: string
          enum: [CANDIDATE, RECRUITER]
        companyName:
          oneOf:
          - { type: string }
          - { type: 'null' }
          description: Nullable and only populated for a Recruiter when a company name exists.
      description: Public Community author projection. It never contains email, resume data, tokens, company verification status, or other private account fields.
...
    CommunityDirectConversationResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [conversationId, participant, createdAt, updatedAt]
          properties:
            conversationId: {type: string}
            participant: {$ref: '#/components/schemas/CommunityAuthor'}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
    CommunityDirectMessage:
      type: object
      required: [messageId, conversationId, senderId, body, sentAt]
      properties:
        messageId: {type: string}
        conversationId: {type: string}
        senderId: {type: string}
        body: {type: string, minLength: 1, maxLength: 2000}
        sentAt: {type: string, format: date-time}
    CommunityDirectMessageListResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: '#/components/schemas/CommunityDirectMessage'}}
        meta: {$ref: '#/components/schemas/PageMeta'}
    CreateCommunityCommentRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 500
          pattern: ^.*\S.*$
          description: Text only. The server trims leading and trailing whitespace before validation; the normalized value must contain 1 to 500 characters and at least one non-whitespace character.
          x-normalization: trim
          x-trimmed-minLength: 1
          x-trimmed-maxLength: 500
...
    CandidateOnboardingRequest:
      type: object
      additionalProperties: false
      required: [headline, location, age, resumeSummary, skills, desiredTitle, preferredLocation, workplaceType, employmentType]
      properties:
        headline: { type: string, minLength: 1, maxLength: 200 }
        location: { type: string, minLength: 1, maxLength: 100 }
        age: { type: integer, minimum: 16, maximum: 100 }
        resumeSummary: { type: string, minLength: 1, maxLength: 5000 }
        skills:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string, minLength: 1, maxLength: 200 }
        desiredTitle: { type: string, minLength: 1, maxLength: 200 }
        preferredLocation: { type: string, minLength: 1, maxLength: 200 }
        workplaceType: { $ref: '#/components/schemas/WorkplaceType' }
        employmentType: { $ref: '#/components/schemas/EmploymentType' }
...
    EmploymentType:
      type: string
      enum:
      - FULL_TIME
      - INTERNSHIP
      - PART_TIME
    WorkplaceType:
      type: string
      enum:
      - ONSITE
      - HYBRID
      - REMOTE
...
    PageMeta:
      type: object
      required:
      - page
      - pageSize
      - total
      - hasNext
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
...
    ErrorDetail:
      type: object
      required:
      - code
      - message
      - requestId
      - fieldErrors
      properties:
        code:
          type: string
          examples:
          - VALIDATION_ERROR
        message:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
...